+++ index.html
@@ -63,7 +63,7 @@
         </main>
 
         <script src="https://cdn.jsdelivr.net/npm/luxon@2.4.0/build/global/luxon.min.js"></script>
-        <script type="module" src="./modules/current_date.js"></script>
+        <script type="module" src="./modules/current-date.js"></script>
         <script type="module" src="index.js"></script>
     </body>
 </html>
