+++ modules/books-manager.js
